refactor(storage): Extract index management into separate module
Extract user-defined index management code from database.rs into a
dedicated indexes.rs module within the database module. This improves
code organization and maintainability.

Changes:
- Create database/indexes.rs with IndexManager, IndexMetadata, and IndexData
- Move database.rs to database/database.rs
- Add database/mod.rs to organize module structure
- Refactor Database struct to use IndexManager internally
- Database methods now delegate to IndexManager for all index operations
- Update storage/lib.rs exports to include new index types

Benefits:
- Reduces database.rs from 701 to 528 lines (~25% reduction)
- Groups related index functionality together (303 lines in indexes.rs)
- Follows single responsibility principle
- Makes index management easier to test in isolation
- Improves code navigation and maintainability

All existing tests pass, including 14 index-specific tests in executor
and 27 storage tests. No breaking changes to public API.

Closes #1005
