+++ crates/storage/src/database/mod.rs
@@ -1,0 +1,9 @@
+// ============================================================================
+// Database Module
+// ============================================================================
+
+mod database;
+pub mod indexes;
+
+pub use database::{Database, Savepoint, TransactionChange, TransactionState};
+pub use indexes::{IndexData, IndexManager, IndexMetadata};
