+++ crates/storage/src/lib.rs
@@ -8,7 +8,7 @@ pub mod error;
 pub mod row;
 pub mod table;
 
-pub use database::{Database, TransactionState};
+pub use database::{Database, IndexData, IndexManager, IndexMetadata, TransactionState};
 pub use error::StorageError;
 pub use row::Row;
 pub use table::Table;
